initial split of ganglia to its own data lib and split of ganglia settings
