+++ zamboni_dashboard/templates/base.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>{% block title %}{{config.DASHBOARD_NAME}}{% endblock %}</title>
+        <title>{% block title %}{{ config.DASHBOARD_NAME }}{% endblock %}</title>
         <link rel="stylesheet" href="{{ url_for('static', filename='bootstrap/css/bootstrap.min.css') }}" />
         <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}" />
         <script src="{{ url_for('static', filename='js/jquery-1.7.2.min.js') }}"></script>
@@ -18,7 +18,7 @@
         <div class="navbar navbar-fixed-top">
             <div class="navbar-inner">
                 <div class="container">
-                    <a class="brand" href="#">{{config.DASHBOARD_NAME}}</a>
+                    <a class="brand" href="#">{{ config.DASHBOARD_NAME }}</a>
                     <ul class="nav">
                         <li><a href="{{ url_for('index') }}">Home</a></li>
                         <li><a href="{{ url_for('ganglia') }}">Ganglia</a></li>
